+++ src/App/Shared/GoldEx.Shared/DTOs/Settings/GetSettingResponse.cs
@@ -5,7 +5,7 @@ public record GetSettingResponse(
     string InstitutionName,
     string Address,
     string PhoneNumber,
-    float TaxPercent,
-    float GoldProfitPercent,
-    float JewelryProfitPercent,
+    decimal TaxPercent,
+    decimal GoldProfitPercent,
+    decimal JewelryProfitPercent,
     TimeSpan PriceUpdateInterval);
