+++ src/App/Shared/GoldEx.Shared/DTOs/Settings/UpdateSettingRequest.cs
@@ -4,7 +4,7 @@ public record UpdateSettingRequest(
     string InstitutionName,
     string Address,
     string PhoneNumber,
-    float TaxPercent,
-    float GoldProfitPercent,
-    float JewelryProfitPercent,
+    decimal TaxPercent,
+    decimal GoldProfitPercent,
+    decimal JewelryProfitPercent,
     TimeSpan PriceUpdateInterval);
